moved MailModal to correct folder, fixed contact buttons and name of map
